Refactor: Input 관련 파일들 위치 이동
input 관련 값들 Input 컴포넌트 안에 위치, Input 컴포넌트 common 폴더 아래에 위치
