--- src/components/Input/index.tsx
@@ -1,1 +1,0 @@
-export { default as Input } from './Input';
