+++ src/components/common/Input/Input.tsx
@@ -1,6 +1,19 @@
 import React from 'react';
 import Styled from './Input.styled';
-import { InputProps } from './types';
+
+export const InputSize = {
+  xs: 'xs',
+  sm: 'sm',
+  md: 'md',
+} as const;
+
+export type InputSizeType = typeof InputSize[keyof typeof InputSize];
+
+export interface InputProps extends React.InputHTMLAttributes<HTMLInputElement> {
+  $size: InputSizeType;
+  label?: string;
+  errorMessage?: string;
+}
 
 const Input = ({ className, $size, label, errorMessage, ...resetProps }: InputProps) => {
   return (
